refactor: pedidos tabrow
