+++ app/src/main/java/com/anotasmart/ui/screens/relatorios/RelatoriosScreen.kt
@@ -16,7 +16,6 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
@@ -25,8 +24,10 @@ import com.anotasmart.AnotaSmartApplication
 import com.anotasmart.ui.components.StandardScreen
 import com.anotasmart.ui.viewModels.RelatoriosViewModel
 import com.anotasmart.ui.viewModels.RelatoriosViewModelFactory
-import com.anotasmart.utils.formatCurrency
-import com.anotasmart.model.PeriodoRelatorio
+import com.anotasmart.ui.screens.relatorios.components.DetailRow
+import com.anotasmart.ui.screens.relatorios.components.PeriodSelector
+import com.anotasmart.ui.screens.relatorios.components.SummaryCard
+import com.anotasmart.utils.formatSafe
 
 
 @Composable
@@ -50,7 +51,8 @@ fun RelatoriosScreen(
         onBackClick = onBackClick,
         hasBottomBar = false
     ) {
-        // Seleção de Período
+
+        // SEÇÃO 1
         item {
             PeriodSelector(
                 selected = periodoSelecionado,
@@ -58,13 +60,37 @@ fun RelatoriosScreen(
             )
         }
 
-        // --- SEÇÃO 1: RESULTADO LÍQUIDO ---
         item {
             SectionHeader("Resumo de Lucros")
         }
-
         item {
-            MainProfitCard(overview.lucroLiquido)
+            Card(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp),
+                colors = CardDefaults.cardColors(
+                    containerColor = if (overview.lucroLiquido >= 0) Color(0xFFE8F5E9) else Color(0xFFFFEBEE)
+                ),
+                shape = RoundedCornerShape(16.dp)
+            ) {
+                Column(
+                    modifier = Modifier.padding(24.dp),
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Text(
+                        text = "Lucro Líquido (Lucro - Despesas)",
+                        style = MaterialTheme.typography.labelLarge,
+                        color = if (overview.lucroLiquido >= 0) Color(0xFF2E7D32) else Color(0xFFC62828)
+                    )
+                    Spacer(modifier = Modifier.height(8.dp))
+                    Text(
+                        text = "R$ ${formatSafe(overview.lucroLiquido)}",
+                        style = MaterialTheme.typography.headlineLarge,
+                        fontWeight = FontWeight.Bold,
+                        color = if (overview.lucroLiquido >= 0) Color(0xFF1B5E20) else Color(0xFFB71C1C)
+                    )
+                }
+            }
         }
 
         item {
@@ -96,23 +122,23 @@ fun RelatoriosScreen(
         item {
             DetailRow(
                 title = "Lucro Estimado (Itens)",
-                value = formatCurrency(overview.lucroEstimado),
+                value = "R$ ${formatSafe(overview.lucroEstimado)}",
                 description = "Ganho real sobre os produtos/serviços",
                 icon = Icons.Default.TrendingUp,
                 color = MaterialTheme.colorScheme.primary
             )
         }
 
-        // --- SEÇÃO 2: FLUXO DE CAIXA ---
+        // SEÇÃO 2
         item {
             Spacer(modifier = Modifier.height(16.dp))
-            SectionHeader("Fluxo de Caixa (Entradas)")
+            SectionHeader("Fluxo de Caixa")
         }
 
         item {
             DetailRow(
                 title = "Dinheiro em Caixa",
-                value = formatCurrency(overview.dinheiroEmCaixa),
+                value = "R$ ${formatSafe(overview.dinheiroEmCaixa)}",
                 description = "Total de parcelas pagas no período",
                 icon = Icons.Default.PriceCheck,
                 color = Color(0xFF2E7D32)
@@ -122,7 +148,7 @@ fun RelatoriosScreen(
         item {
             DetailRow(
                 title = "Contas a Receber",
-                value = formatCurrency(overview.contasAReceber),
+                value = "R$ ${formatSafe(overview.contasAReceber)}",
                 description = "Tudo que ainda falta receber (total)",
                 icon = Icons.Default.AccountBalanceWallet,
                 color = MaterialTheme.colorScheme.secondary
@@ -132,7 +158,7 @@ fun RelatoriosScreen(
         item {
             DetailRow(
                 title = "Inadimplência",
-                value = formatCurrency(overview.inadimplencia),
+                value = "R$ ${formatSafe(overview.inadimplencia)}",
                 description = "Parcelas vencidas e não pagas",
                 icon = Icons.Default.Warning,
                 color = Color(0xFFE65100)
@@ -156,7 +182,24 @@ fun RelatoriosScreen(
 
         item {
             Spacer(modifier = Modifier.height(24.dp))
-            InfoExplanationCard()
+            Surface(
+                color = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f),
+                shape = RoundedCornerShape(12.dp)
+            ) {
+                Column(modifier = Modifier.padding(16.dp)) {
+                    Text(
+                        text = "Entenda os conceitos:",
+                        style = MaterialTheme.typography.labelLarge,
+                        fontWeight = FontWeight.Bold,
+                        color = MaterialTheme.colorScheme.primary
+                    )
+                    Spacer(modifier = Modifier.height(8.dp))
+
+                    ExplanationItem("Lucro Líquido", "É o lucro das vendas menos as despesas gerais.")
+                    ExplanationItem("Dinheiro em Caixa", "Refere-se apenas ao que já foi efetivamente pago no período.")
+                    ExplanationItem("Inadimplência", "Soma de todas as parcelas que já passaram da data de vencimento.")
+                }
+            }
         }
     }
 }
@@ -173,35 +216,6 @@ fun SectionHeader(title: String) {
 }
 
 @Composable
-fun InfoExplanationCard() {
-    Surface(
-        color = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f),
-        shape = RoundedCornerShape(12.dp)
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            Text(
-                text = "Entenda os conceitos:",
-                style = MaterialTheme.typography.labelLarge,
-                fontWeight = FontWeight.Bold,
-                color = MaterialTheme.colorScheme.primary
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            
-            ExplanationItem("Lucro Líquido", "É o lucro das vendas menos as despesas gerais.")
-            ExplanationItem("Dinheiro em Caixa", "Refere-se apenas ao que já foi efetivamente pago no período.")
-            ExplanationItem("Inadimplência", "Soma de todas as parcelas que já passaram da data de vencimento.")
-            
-            Spacer(modifier = Modifier.height(8.dp))
-            Text(
-                text = "Nota: O lucro das vendas usa o preço de custo registrado no ato da venda.",
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant
-            )
-        }
-    }
-}
-
-@Composable
 fun ExplanationItem(label: String, text: String) {
     Column(modifier = Modifier.padding(vertical = 4.dp)) {
         Text(text = label, style = MaterialTheme.typography.labelMedium, fontWeight = FontWeight.Bold)
@@ -209,146 +223,3 @@ fun ExplanationItem(label: String, text: String) {
     }
 }
 
-@Composable
-fun PeriodSelector(
-    selected: PeriodoRelatorio,
-    onSelected: (PeriodoRelatorio) -> Unit
-) {
-    PrimaryScrollableTabRow(
-        selectedTabIndex = selected.ordinal,
-        edgePadding = 16.dp,
-        containerColor = MaterialTheme.colorScheme.surface,
-        contentColor = MaterialTheme.colorScheme.primary,
-        divider = {}
-    ) {
-        PeriodoRelatorio.entries.forEach { periodo ->
-            val isSelected = selected == periodo
-            Tab(
-                selected = isSelected,
-                onClick = { onSelected(periodo) },
-                text = {
-                    Text(
-                        text = when (periodo) {
-                            PeriodoRelatorio.HOJE -> "Hoje"
-                            PeriodoRelatorio.SEMANA -> "Semana"
-                            PeriodoRelatorio.MES -> "Mês"
-                            PeriodoRelatorio.ANO -> "Ano"
-                            PeriodoRelatorio.TUDO -> "Tudo"
-                        },
-                        style = MaterialTheme.typography.labelLarge,
-                        fontWeight = if (isSelected) FontWeight.Bold else FontWeight.Normal
-                    )
-                },
-                unselectedContentColor = MaterialTheme.colorScheme.onSurfaceVariant,
-                selectedContentColor = MaterialTheme.colorScheme.primary
-            )
-        }
-    }
-}
-
-@Composable
-fun MainProfitCard(lucro: Double) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 8.dp),
-        colors = CardDefaults.cardColors(
-            containerColor = if (lucro >= 0) Color(0xFFE8F5E9) else Color(0xFFFFEBEE)
-        ),
-        shape = RoundedCornerShape(16.dp)
-    ) {
-        Column(
-            modifier = Modifier.padding(24.dp),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = "Lucro Líquido no Período",
-                style = MaterialTheme.typography.labelLarge,
-                color = if (lucro >= 0) Color(0xFF2E7D32) else Color(0xFFC62828)
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            Text(
-                text = formatCurrency(lucro),
-                style = MaterialTheme.typography.headlineLarge,
-                fontWeight = FontWeight.Bold,
-                color = if (lucro >= 0) Color(0xFF1B5E20) else Color(0xFFB71C1C)
-            )
-        }
-    }
-}
-
-@Composable
-fun SummaryCard(
-    title: String,
-    value: Double,
-    description: String,
-    icon: ImageVector,
-    color: Color,
-    modifier: Modifier = Modifier
-) {
-    Card(
-        modifier = modifier,
-        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-        shape = RoundedCornerShape(12.dp)
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            Icon(imageVector = icon, contentDescription = null, tint = color, modifier = Modifier.size(20.dp))
-            Spacer(modifier = Modifier.height(12.dp))
-            Text(text = title, style = MaterialTheme.typography.labelMedium, color = MaterialTheme.colorScheme.onSurfaceVariant)
-            Text(
-                text = formatCurrency(value),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Bold,
-                color = MaterialTheme.colorScheme.onSurface
-            )
-            Text(
-                text = description,
-                style = MaterialTheme.typography.labelSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.7f)
-            )
-        }
-    }
-}
-
-@Composable
-fun DetailRow(
-    title: String,
-    value: String,
-    description: String,
-    icon: ImageVector,
-    color: Color
-) {
-    Surface(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 6.dp),
-        color = MaterialTheme.colorScheme.surface,
-        shape = RoundedCornerShape(12.dp),
-        tonalElevation = 1.dp
-    ) {
-        Row(
-            modifier = Modifier.padding(16.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(40.dp)
-                    .background(color.copy(alpha = 0.1f), RoundedCornerShape(8.dp)),
-                contentAlignment = Alignment.Center
-            ) {
-                Icon(imageVector = icon, contentDescription = null, tint = color, modifier = Modifier.size(24.dp))
-            }
-            Spacer(modifier = Modifier.width(16.dp))
-            Column(modifier = Modifier.weight(1f)) {
-                Text(text = title, style = MaterialTheme.typography.bodyMedium, color = MaterialTheme.colorScheme.onSurfaceVariant)
-                Text(text = value, style = MaterialTheme.typography.titleMedium, fontWeight = FontWeight.Bold)
-                Text(
-                    text = description,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-        }
-    }
-}
